production/retriever-server/requirements.txt
farm-haystack==0.10.0
python-decouple==3.4
flask==1.1.2
waitress==2.0.0
